Add files via upload
I have only tested this on one account so I am not sure if Facebook has any account-specific fields that would need to be altered for the code to work if you could let me know that would be amazing. Also please note that I am not an expert with selenium so if anyone knows any ways to improve the code to make it operate better or more efficiently then please let me know.

I suggest to change some of the delays to make it more suitable for your computer and internet speeds.