GUI version with save functionality
This is a more user-friendly version with a GUI made using Tkinter and has a save feature that saves and loads one set of credentials to and from a JSON file. For the save functionality to work properly you need to enter the credentials into the entry boxes first and then click the start button. To load the last saved credentials you need to leave the entry boxes blank and click the load credentials checkbox and then click the start button.